# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dev Event API",
        "title": "Dev Event Server",
        "contact": {},
        "license": {}
    },
    "paths": {
        "/admin/events": {
            "post": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "새 이벤트 생성",
                "tags": [
                    "Admin - Events"
                ],
                "parameters": [
                    {
                        "description": "create event",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Event"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "400": {},
                    "401": {},
                    "500": {}
                }
            }
        },
        "/admin/events/:eventID": {
            "get": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "이벤트 정보 조회",
                "tags": [
                    "Admin - Events"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "eventID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            },
            "put": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "이벤트 수정",
                "tags": [
                    "Admin - Events"
                ],
                "parameters": [
                    {
                        "description": "Save event",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Event"
                        }
                    },
                    {
                        "type": "string",
                        "description": "eventID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            },
            "delete": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "이벤트 삭제",
                "tags": [
                    "Admin - Events"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "eventID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            }
        },
        "/admin/tags": {
            "get": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "모든 태그 조회",
                "tags": [
                    "Admin - Tags"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Tag"
                            }
                        }
                    },
                    "400": {},
                    "401": {},
                    "500": {}
                }
            },
            "post": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "태그 생성",
                "tags": [
                    "Admin - Tags"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "tagID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            },
            "delete": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "태그 삭제",
                "tags": [
                    "Admin - Tags"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "tagID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Message"
                            }
                        }
                    },
                    "400": {},
                    "401": {},
                    "500": {}
                }
            }
        },
        "/admin/tags/:tagID": {
            "put": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "태그 수정",
                "tags": [
                    "Admin - Tags"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "tagID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "Get all users info",
                "tags": [
                    "Admin - Users"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Users"
                            }
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            }
        },
        "/admin/users/:userID": {
            "get": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "회원 정보 조회",
                "tags": [
                    "Admin - Users"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Users"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            },
            "put": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "회원 정보 수정",
                "tags": [
                    "Admin - Users"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Users"
                        }
                    },
                    "400": {},
                    "401": {},
                    "404": {},
                    "500": {}
                }
            },
            "delete": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "회원 정보 삭제",
                "tags": [
                    "Admin - Users"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "로그인",
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "description": "로그인",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/requests.Login"
                        }
                    },
                    "401": {},
                    "500": {}
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "회원가입",
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "description": "register user",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "400": {},
                    "500": {}
                }
            }
        },
        "/events/:year/:month": {
            "get": {
                "description": "이벤트 가져오기 (년, 월)",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Event"
                            }
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "나의 프로필 확인",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Users"
                        }
                    },
                    "401": {},
                    "500": {}
                }
            },
            "put": {
                "security": [
                    {
                        "userAPIKey": []
                    }
                ],
                "description": "프로필 수정",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Message"
                        }
                    },
                    "401": {},
                    "404": {},
                    "500": {}
                }
            }
        }
    },
    "definitions": {
        "models.Event": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "day": {
                    "type": "integer"
                },
                "free": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imgURL": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "tages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "requests.Event": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "day": {
                    "type": "integer"
                },
                "free": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imgURL": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "requests.Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "requests.RegisterUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "responses.Event": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "day": {
                    "type": "integer"
                },
                "free": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imgURL": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "tages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.Tag"
                    }
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "responses.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Tag": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "responses.Users": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}